Update to Tauri v2 beta
