+++ src-tauri/Cargo.toml
@@ -9,13 +9,22 @@ default-run = "gale"
 edition = "2021"
 rust-version = "1.79"
 
+[lib]
+name = "gale_lib"
+path = "src/lib.rs"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
+[[bin]]
+name = "gale"
+path = "src/main.rs"
+
 [build-dependencies]
-tauri-build = { version = "1.5.1", features = [] }
+tauri-build = { version = "2.0.0-beta", features = [] }
 
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.6.7", features = [ "fs-read-file", "app-all", "updater", "clipboard-all", "shell-open", "http-request", "window-all", "dialog-all", "clipboard" ] }
+tauri = { version = "2.0.0-beta", features = ["unstable"] }
 tokio = { version = "1", features = ["full"] }
 
 anyhow = "1.0"
@@ -32,7 +41,7 @@ reqwest =  { version = "0.12", features = ["json", "stream"] }
 walkdir = "2"
 typeshare = "1.0"
 image = "0.25"
-semver = "1.0"
+semver = { version = "1.0", features = ["serde"] }
 serde_yaml = "0.9"
 base64 = "0.22"
 lazy_static = "1.4"
@@ -40,8 +49,15 @@ heck = "0.5"
 log = "0.4"
 simplelog = "0.12"
 chrono = { version = "0.4", features = ["serde"] }
-tauri-plugin-deep-link = "=0.1.2"
-tauri-plugin-window-state = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
+tauri-plugin-deep-link = "2.0.0-beta.9"
+tauri-plugin-window-state = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v2" }
+tauri-plugin-dialog = "2.0.0-beta.11"
+tauri-plugin-shell = "2.0.0-beta.9"
+tauri-plugin-os = "2.0.0-beta.8"
+tauri-plugin-process = "2.0.0-beta.8"
+tauri-plugin-updater = "2.0.0-beta.10"
+tauri-plugin-clipboard-manager = "2.1.0-beta.6"
+tauri-plugin-http = "2.0.0-beta.12"
 keyring = "2.3"
 dirs-next = "2.0.0"
 
@@ -49,7 +65,7 @@ dirs-next = "2.0.0"
 objc = "0.2"
 
 [target.'cfg(target_os="linux")'.dependencies]
-webkit2gtk = "0.18"
+webkit2gtk = "2.0.1"
 
 [features]
 custom-protocol = [ "tauri/custom-protocol" ]
