+++ src-tauri/capabilities/main.json
@@ -1,0 +1,25 @@
+{
+  "identifier": "main-capability",
+  "description": "Capability for the main window",
+  "windows": ["main"],
+  "permissions": [
+    "path:default",
+    "event:default",
+    "window:default",
+    "app:default",
+    "resources:default",
+    "menu:default",
+    "tray:default",
+    "deep-link:default",
+    "http:allow-fetch",
+    "http:default",
+    "updater:allow-check",
+    "updater:default",
+    "dialog:default",
+    {
+      "identifier": "http:default",
+      "allow": [{ "url": "https://thunderstore.io/api/*" }, { "url": "https://raw.githubusercontent.com/Kesomannen/gale/master/CHANGELOG.md" }],
+      "deny": []
+    }
+  ]
+}
