+++ src-tauri/src/util/window.rs
@@ -2,12 +2,12 @@ pub trait WindowExt {
     fn zoom(&self, factor: f64) -> tauri::Result<()>;
 }
 
-impl WindowExt for tauri::Window {
+impl WindowExt for tauri::WebviewWindow {
     fn zoom(&self, scale_factor: f64) -> tauri::Result<()> {
         self.with_webview(move |webview| {
             #[cfg(target_os = "linux")]
             {
-                use webkit2gtk::traits::WebViewExt;
+                use webkit2gtk::WebViewExt;
                 webview.inner().set_zoom_level(scale_factor);
             }
